[LEX] Store a ref to src instead of duplicating the text
This simply change allow us to shrink the size of a single Token from 72
bytes to 32 bytes.

We could decrease the size even further to 16 bytes by exploiting the source
location range. This does not come from free, a method taking in a SourceManager
is required. This hurts the usability but decreases memory consumption.