[HIR] Start using SimbolicScope and add Symbol to all decl
Every declaration creates a symbol that could be referenced later. Every
Symbol has a hir::Type, a kind and a name. This name, given current language
needs, is mandatory but could be optional.

Every Symbol lies in a SymbolicScope. A SymbolicScope is a logical place
which it is created per language constructs. SymbolicScopes are nested
within other SymbolicScopes, except for the Global SymbolicScope which
is created for the CompilationUnit and is the top most.

For now on, every Decl node has a Symbol. Also, HIRBuilder has a stack
of SymbolicScope which is used to determine if a Symbol is already
declared for the current SymbolicScope.

All Symbols and all SimbolicScopes are allocated in an arena in the
HIRContext.